Moved `namespaces.h` away from the API.
